# Conflicts:
#   GroupProjectApi/.idea/dataSources.xml
#   GroupProjectApi/src/main/java/s3_gps_ivanti/business/application/impl/CreateApplicationUseCaseImpl.java
#   gps-react-app/src/components/Header.js
#   gps-react-app/src/pages/ContentCreator/ApplicationPage/ApplicationPage.js
#   gps-react-app/src/pages/ContentCreator/MyAppsPage/MyAppsPage.js
#   gps-react-app/src/pages/ContentCreator/UpdateApplicationPage.js

+++ GroupProjectApi/src/main/java/s3_gps_ivanti/dto/application/ApplicationDetailedInfoDTO.java
@@ -21,7 +21,7 @@ public class ApplicationDetailedInfoDTO {
     private String description;
     private List<String> screenshots;
     private int totalDownloads;
-    private int avgRating;
+    private double avgRating;
     private List<VersionDTO> versions;
     private List<ReviewDTO> reviews;
     private CustomerSmallDetailDTO creator;
